feat: enum에서 const 객체 사용
